+++ packages/sdk/src/index.e2e.test.ts
@@ -5,6 +5,7 @@ import { MermaidChart } from './index.js';
 import { beforeAll, describe, expect, it } from 'vitest';
 
 import process from 'node:process';
+import { AxiosError } from 'axios';
 
 let client: MermaidChart;
 
@@ -30,5 +31,45 @@ describe('getUser', () => {
     const user = await client.getUser();
 
     expect(user).toHaveProperty('emailAddress');
+  });
+});
+
+const documentMatcher = expect.objectContaining({
+  documentID: expect.any(String),
+  major: expect.any(Number),
+  minor: expect.any(Number),
+});
+
+describe("getDocument", () => {
+  it("should get publicly shared diagram", async() => {
+    const latestDocument = await client.getDocument({
+      // owned by alois@mermaidchart.com
+      documentID: '8bce727b-69b7-4f6e-a434-d578e2b363ff',
+    });
+
+    expect(latestDocument).toStrictEqual(documentMatcher);
+
+    const earliestDocument = await client.getDocument({
+      // owned by alois@mermaidchart.com
+      documentID: '8bce727b-69b7-4f6e-a434-d578e2b363ff',
+      major: 0,
+      minor: 1,
+    });
+
+    expect(earliestDocument).toStrictEqual(documentMatcher);
+  });
+
+  it("should throw 404 on unknown document", async() => {
+    let error: AxiosError | undefined = undefined;
+    try {
+      await client.getDocument({
+        documentID: '00000000-0000-0000-0000-0000deaddead',
+      });
+    } catch (err) {
+      error = err as AxiosError;
+    }
+
+    expect(error).toBeInstanceOf(AxiosError);
+    expect(error?.response?.status).toBe(404);
   });
 });